src/test/java/guru/springframework/services/UnitOfMeasureServiceImplTest.java
package guru.springframework.services;

import guru.springframework.commands.UnitOfMeasureCommand;
import guru.springframework.converters.UnitOfMeasureToUnitOfMeasureCommand;
import guru.springframework.domain.UnitOfMeasure;
import guru.springframework.repository.UnitOfMeasureRepository;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.Mock;

import java.util.HashSet;
import java.util.Set;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.mockito.Mockito.*;
import static org.mockito.MockitoAnnotations.initMocks;

class UnitOfMeasureServiceImplTest {

    @Mock
    UnitOfMeasureRepository unitOfMeasureRepository;

    UnitOfMeasureService unitOfMeasureService;

    @BeforeEach
    void setUp() {
        initMocks(this);
        unitOfMeasureService = new UnitOfMeasureServiceImpl(unitOfMeasureRepository,
                new UnitOfMeasureToUnitOfMeasureCommand());
    }

    @Test
    void listAllUoms() {
        UnitOfMeasure unitOfMeasure1 = new UnitOfMeasure();
        unitOfMeasure1.setId(1L);
        unitOfMeasure1.setDescription("tablespoon");

        UnitOfMeasure unitOfMeasure2 = new UnitOfMeasure();
        unitOfMeasure2.setId(2L);
        unitOfMeasure2.setDescription("teaspoon");

        Set<UnitOfMeasure> iterable = new HashSet<>();
        iterable.add(unitOfMeasure1);
        iterable.add(unitOfMeasure2);

        when(unitOfMeasureRepository.findAll()).thenReturn(iterable);

        Set<UnitOfMeasureCommand> set = unitOfMeasureService.listAllUoms();

        assertEquals(iterable.size(), set.size());
        assertEquals(2, set.size());
        assertNotNull(set);
        verify(unitOfMeasureRepository, times(1)).findAll();
    }
}
